+++ NEWS.md
@@ -1,5 +1,6 @@
 # Changes in v0.5.2
 
+* Added predict() to identify topics of unseen documents (#9)
 * Allow selecting seed words based on their frequencies using dfm_trim() in textmodel_seededlda() via ... (#8)
 
 # Changes in v0.5.1